move weapon to parent directory + bug fix
When the player shoots, the game pauses and resumes when the bullet has
finished bouncing around.
But if the player shoots and its gun isn't ready yet
(mainly because of the weapon's cooldown), no bullet is generated
but the game pauses nontheless and stays in that state for ever.
fix: pause the game only if the gun's realy shoot